Code/Weather API/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
gunicorn==21.2.0
python-decouple==3.8
pytz==2022.6
sqlparse==0.4.3
requests==2.31.0